Refactor hello example
