+++ .gitignore
@@ -23,11 +23,14 @@ _testmain.go
 *.prof
 
 # Executables
-examples/hello/hello
+cmd/goapp/goapp
+examples/hello/bin/hello-mac/hello-mac
+examples/hello/bin/hello-web/hello-web
 examples/menu/menu
 examples/dock/dock
 examples/nav/nav
 examples/dragdrop/dragdrop
 examples/test/test
 *.exe
+*/goapp.js
