+++ examples/hello/bin/hello-web/main.go
@@ -1,0 +1,15 @@
+package main
+
+import (
+	"github.com/murlokswarm/app"
+	"github.com/murlokswarm/app/drivers/web"
+	"github.com/murlokswarm/app/examples/hello"
+)
+
+func main() {
+	app.Import(&hello.Hello{})
+
+	app.Run(&web.Driver{
+		DefaultURL: "/hello.Hello",
+	})
+}
