+++ examples/nav/bin/nav-mac/main.go
@@ -1,0 +1,34 @@
+package main
+
+import (
+	"github.com/murlokswarm/app"
+	"github.com/murlokswarm/app/drivers/mac"
+	"github.com/murlokswarm/app/examples/nav"
+)
+
+func main() {
+	app.Import(&nav.City{})
+
+	app.Run(&mac.Driver{
+		OnRun: func() {
+			newWindow()
+		},
+
+		OnReopen: func(hasVisibleWindow bool) {
+			if !hasVisibleWindow {
+				newWindow()
+			}
+		},
+	})
+}
+
+func newWindow() {
+	app.NewWindow(app.WindowConfig{
+		Title:           "nav",
+		TitlebarHidden:  true,
+		Width:           1280,
+		Height:          768,
+		BackgroundColor: "#21252b",
+		DefaultURL:      "/nav.City",
+	})
+}
