+++ examples/nav/bin/nav-web/main.go
@@ -1,0 +1,15 @@
+package main
+
+import (
+	"github.com/murlokswarm/app"
+	"github.com/murlokswarm/app/drivers/web"
+	"github.com/murlokswarm/app/examples/nav"
+)
+
+func main() {
+	app.Import(&nav.City{})
+
+	app.Run(&web.Driver{
+		DefaultURL: "/nav.City",
+	})
+}
