--- examples/nav/main_darwin_amd64.go
@@ -1,31 +1,0 @@
-package main
-
-import (
-	"github.com/murlokswarm/app"
-	"github.com/murlokswarm/app/drivers/mac"
-)
-
-func main() {
-	app.Run(&mac.Driver{
-		OnRun: func() {
-			newWindow()
-		},
-
-		OnReopen: func(hasVisibleWindow bool) {
-			if !hasVisibleWindow {
-				newWindow()
-			}
-		},
-	})
-}
-
-func newWindow() {
-	app.NewWindow(app.WindowConfig{
-		Title:           "nav",
-		TitlebarHidden:  true,
-		Width:           1280,
-		Height:          768,
-		BackgroundColor: "#21252b",
-		DefaultURL:      "/City",
-	})
-}
